+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.2+1 - 2023-05-06
+
+### Added
+
+- DefaultSettingsValue that returns non-nullable values
+- `reset()` method to reset a SettingsValue to its `defaultValue`.
+
+### Changed
+
+- Improved readme and docs
+
 ## 0.1.1+1 - 2023-05-02
 
 ### Fixed